A few fixes to the client application tests.
One thing that is still a problem is that the Server produces a strange
message when a client disconnects. In the console it shows as many empty
spaces followed by the character ' and 2 log entries "Invalid message
received from client" and "Error! Connection lost!". This does not
affect functionality of puts / gets and the server continues to operate
normally. Please note I'm still using the server from BasicStorageServer
(the ms2-server.jar), since there doesn't seem to be a way of running
the new Server application yet.